changed broadcast id type to int

+++ folders.go
@@ -290,7 +290,7 @@ func (f *folders) folderCreate(cd ConnData, data []byte) (string, error) {
 	newName := addFolderTo(parent.Folders, name, newFolder())
 	f.saveFolders()
 	dir = dir[:len(dir)-len(name)] + newName
-	bid := int64(broadcastImageFolderAdd)
+	bid := broadcastImageFolderAdd
 	if f.fileType == fileTypeAudio {
 		bid -= 1
 	} else if f.fileType == fileTypeCharacter {
@@ -333,7 +333,7 @@ func (f *folders) itemMove(cd ConnData, data []byte) (string, error) {
 	newName = addItemTo(newParent.Items, newName, iid)
 	f.saveFolders()
 	itemMove.To += "/" + newName
-	bid := int64(broadcastImageItemMove)
+	bid := broadcastImageItemMove
 	if f.fileType == fileTypeAudio {
 		bid -= 1
 	} else if f.fileType == fileTypeCharacter {
@@ -371,7 +371,7 @@ func (f *folders) folderMove(cd ConnData, data []byte) (string, error) {
 	newName = addFolderTo(newParent.Folders, newName, fd)
 	f.saveFolders()
 	folderMove.To += "/" + newName
-	bid := int64(broadcastImageFolderMove)
+	bid := broadcastImageFolderMove
 	if f.fileType == fileTypeAudio {
 		bid -= 1
 	} else if f.fileType == fileTypeCharacter {
@@ -395,7 +395,7 @@ func (f *folders) itemDelete(cd ConnData, data []byte) error {
 	delete(parent.Items, oldName)
 	f.unlink(iid)
 	f.saveFolders()
-	bid := int64(broadcastImageItemRemove)
+	bid := broadcastImageItemRemove
 	if f.fileType == fileTypeAudio {
 		bid -= 1
 	} else if f.fileType == fileTypeCharacter {
@@ -433,7 +433,7 @@ func (f *folders) folderDelete(cd ConnData, data []byte) error {
 		}
 	})
 	f.saveFolders()
-	bid := int64(broadcastImageFolderRemove)
+	bid := broadcastImageFolderRemove
 	if f.fileType == fileTypeAudio {
 		bid -= 1
 	} else if f.fileType == fileTypeCharacter {
@@ -463,7 +463,7 @@ func (f *folders) linkItem(cd ConnData, data []byte) (string, error) {
 	newName := addItemTo(parent.Items, name, link.ID)
 	f.saveFolders()
 	link.Name = link.Name[:len(link.Name)-len(name)] + newName
-	bid := int64(broadcastImageItemLink)
+	bid := broadcastImageItemLink
 	if f.fileType == fileTypeAudio {
 		bid -= 1
 	} else if f.fileType == fileTypeCharacter {
